Refactoring from class-based to functional components, renamed slice.js to productSlice.js, and currently refactoring Products.js into a functional component.
